requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
rdkit>=2023.3.2
cairosvg>=2.7.0
